+++ common/src/main/java/net/mehvahdjukaar/polytone/sound/PolytoneSoundType.java
@@ -130,7 +130,7 @@ public class PolytoneSoundType extends SoundType {
     public static final Codec<SoundType> REFERENCE_OR_COPY_CODEC = Codec.STRING.flatXmap(s -> {
                 if (s.startsWith("copy(")) {
                     String target = s.replace("copy(", "").replace(")", "");
-                    ResourceLocation r = ResourceLocation.parse(target);
+                    ResourceLocation r = ResourceLocation.tryParse(target);
                     if (r == null) {
                         return DataResult.error(() -> "Invalid string for Sound Type Copy function: " + s + ". Expected 'copy([some_mod]:[some_block])'");
                     }
@@ -141,9 +141,9 @@ public class PolytoneSoundType extends SoundType {
                 }
                 var vanilla = SOUND_NAMES.get(s);
                 if (vanilla != null) return DataResult.success(vanilla);
-                ResourceLocation r = ResourceLocation.parse(s);
+                ResourceLocation r = ResourceLocation.tryParse(s);
                 if (r != null) {
-                    var custom = Polytone.SOUND_TYPES.getCustomSoundType(ResourceLocation.parse(s));
+                    var custom = Polytone.SOUND_TYPES.getCustomSoundType(new ResourceLocation(s));
                     if (custom != null) return DataResult.success(custom);
                 }
                 return DataResult.error(() -> "Could not find any custom Sound Type with id " + r +