+++ changelog.md
@@ -1,3 +1,2 @@
-backported even more changes
-fixed many instances of wrong fields content failing silently without throwing an eror as expected
-fixed custom sound types being broken
+- backported some fixes
+- fixed a crash on fabric due to custom particles